Move seeds table to KnowTarget.cpp
